(`c2rust-analyze`) Moved `IsTrivial` and its `impl`s into its own `mod trivial`.
